+++ lib/data_layer.ex
@@ -2452,6 +2452,9 @@ defmodule AshPostgres.DataLayer do
 
       {:error, error} ->
         {:error, error}
+
+      {:error, :no_rollback, error} ->
+        {:error, :no_rollback, error}
     end
   end